refactor: move placeholder display to own file
